% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFourWay.R
\name{ggFourWay}
\alias{ggFourWay}
\title{Four-Way plot for comparison of log fold changes in a multiple factor RNA seq
experiment}
\usage{
ggFourWay(data, x.level, y.level, control, factor, padj = 0.05, xlim = NULL,
  ylim = NULL, lfc = NULL)
}
\arguments{
\item{data}{a DESeq object class with multiple factors and levels}

\item{x.level}{treatment for the x-axis of the plot}

\item{y.level}{treatment for the y-axis of the plot}

\item{control}{experimental control used in the experiment}

\item{factor}{factor given for the DESeq object for specific treatments}

\item{padj}{significance threshold for adjusted p-value highlighting}

\item{xlim}{optional limits to the x-axis}

\item{ylim}{optional limits to the y-axis}

\item{lfc}{foldchange threshold for visualization}
}
\description{
This function allows you to extract necessary results-based data from a 
DESeq object class to create a four-way plot to compare log fold changes in 
various treatments using ggplot2 aesthetics.
}
\details{
This function allows the user to extract various elements from a DESeq object
class which in turn, creates a temporary data frame to plot the necessary
ggplot aesthetics. In order for this function to work, RNA seq experiments
must have multiple factors and levels including treatments and controls.
By having the recommended criteria, this function will extract the necessary
data by exploiting DESeq's 'contrast' parameters. Data points with 'extreme' 
values that exceed the default viewing frame of the plot will change character 
classes (i.e. points of interest a substantially large log fold change).
}
\examples{
ggFourWay()
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
