% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMA.R
\name{ggMA}
\alias{ggMA}
\title{MA plot from mean expression and log fold changes}
\usage{
ggMA(data, ylim = NULL, padj = 0.05, lfc = 1)
}
\arguments{
\item{data}{a DESeq object class}

\item{ylim}{optional limits to the y axis}

\item{padj}{significance threshold for adjusted p-value highlighting}

\item{lfc}{fold change threshold for visualization}
}
\description{
This function allows you to extract necessary results-based data from a 
DESeq object class to create a MA plot (i.e. a scatter plot) of log2 fold 
changes versus normalized mean counts while implementing ggplot2 aesthetics.
}
\details{
This function allows the user to extract various elements from a DESeq object
class which in turn, creates a temporary data frame to plot the necessary
ggplot aesthetics. Data points with 'extreme' values that exceed the default
viewing frame of the plot will change character classes (i.e. points of 
interest a substantially large log fold change).
}
\examples{
ggMA()
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
