% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{ggVolcano}
\alias{ggVolcano}
\title{Volcano plot from log fold changes and -log10 of the p-value}
\usage{
ggVolcano(data, xlim = NULL, padj = 0.05)
}
\arguments{
\item{data}{a DESeq object class}

\item{xlim}{optional limits to the x axis}

\item{padj}{significance threshold for adjusted p-value highlighting}
}
\description{
This function allows you to extract necessary results-based data from a 
DESEq object class to create a volcano plot (i.e. a scatter plot) of the
negative log of the p-value versus the log of the fold change while
implementing ggplot2 aesthetics.
}
\details{
This function allows the user to extract various elements from a DESeq object
class which in turn, creates a temporary data frame to plot the necessary
ggplot aesthetics. Data points with 'extreme' values that exceed the default
viewing frame of the plot will change character classes (i.e. points of 
interest that either have a substantially large -log10 p-value or log fold
change).
}
\examples{
ggVolcano()
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
